pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "licenseid"
version = "0.1.0"
description = "SPDX License ID matcher with SQLite and Hybrid Search"
readme = "README.md"
requires-python = ">=3.10"
license = "Apache-2.0"
authors = [{ name = "Arthit Suriyawongkul", email = "suriyawa@tcd.ie" }]
keywords = [
    "license",
    "matcher",
    "spdx",
    "license-matcher",
    "opensource",
    "licenseid",
]
classifiers = [
    "Development Status :: 3 - Alpha",
    "Intended Audience :: Developers",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
]
dependencies = [
    "beautifulsoup4>=4.10.0",
    "click>=8.0.0",
    "rapidfuzz>=3.0.0",
    "requests>=2.28.0",
]

[project.optional-dependencies]
dev = ["flake8", "mypy", "pylint", "pytest", "pytest-cov", "ruff"]
java = ["jpype1>=1.7.0"]

[project.scripts]
licenseid = "licenseid.cli:main"

[tool.hatch.build.targets.wheel]
packages = ["src/licenseid"]

[tool.ruff]
line-length = 88
target-version = "py310"

[tool.mypy]
strict = true
python_version = "3.10"

[tool.pytest.ini_options]
markers = [
    "benchmark: marks tests as slow benchmark (deselect with '-m \"not benchmark\"')",
]
